backend/requirements.txt
# Django框架
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# AI服务
google-generativeai>=0.3.0

# 文件处理
PyPDF2>=3.0.0
python-docx>=1.1.0

# 数据处理
pandas>=1.5.0
numpy>=1.24.0

# 环境变量管理
python-dotenv>=1.0.0

# 工具库
werkzeug>=2.3.0

# 可选：高级AI功能（如果需要使用现有的ai_services模块）
# langchain>=0.1.0
# langchain-community>=0.0.20
# chromadb>=0.4.0
# sentence-transformers>=2.2.0
# transformers>=4.30.0
